requirements.txt
asgiref==3.5.0
autopep8==1.5.7
certifi==2021.10.8
Django==3.2.12
django-phonenumber-field==6.1.0
phonenumberslite==8.12.45
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
